feat: add global styles and new components for the frontend
- Import global CSS in frontend.tsx
- Create Footer component for consistent footer layout
- Implement AdminPage with header and footer
- Add DashboardShell for managing DNS zones and records
- Create HomePage with features and documentation sections
- Introduce api.ts for API request handling and type definitions
