+++ Frontend/src/frontend.tsx
@@ -6,6 +6,7 @@
  */
 
 import { createRoot } from "react-dom/client";
+import "./global.css";
 import { App } from "./App";
 
 function start() {
