+++ Frontend/tsconfig.json
@@ -7,6 +7,8 @@
     "moduleDetection": "force",
     "jsx": "react-jsx",
     "allowJs": true,
+    "baseUrl": ".",
+    "ignoreDeprecations": "6.0",
 
     // Bundler mode
     "moduleResolution": "bundler",
